Initial commit
This is an initial proof of concept.

`transform` expects some `data` which is either a string or bytes, a
`provider` for a transformer (encoder or decoder as appropriate), and
keyword arguments for the provider. It instantiates the transformer with
the keyword arguments and uses it to transform the data, by repeatedly
calling it and concatenating the resulting `chunks`.

A transformer is a function which accepts data and returns a 2-tuple of
transformed data and the amount of data transformed. It may only
partially transform the data *even if* it could transform more, and it
may return non-empty data while reporting that no data was transformed.
However, whenever no data is transformed, this is a signal that no
future calls will succeed.

When `full==True` for `transform` (the default), it insists that the
transformer transforms all supplied data (potentially across multiple
calls), or else a `ValueError` is raised. Otherwise, the partially
transformed and concatenated data is returned.

Future plans:

* It should be possible for `data` to be a file-like object
  (recognizably open in either a text mode or binary mode). In this
  case, a wrapper file-like object is returned.
* Type annotations should be used to detect the input and output types
  (bytes or str) of the transformer.
* A provider should be able to provide both the "encoder" and "decoder"
  for a bytes<->str transformation, and the appropriate one should be
  chosen according to the data (or method of a stream wrapper).
* Chaining stream wrappers should work transparently.
* There should be an object representing a codec - essentially binding a
  provider to `transform`, so that when called it does the right thing.
  The `full` argument should be removed from the interface as an
  implementation detail - essentially it should be fixed `True` when the
  innermost `data` is a bytes or str, and `False` when it's a stream (in
  most cases, at least).

In short: it should be possible to have uses like

with foo_codec(bar_codec(baz_codec(open('foo')))) as stream:
    print(stream.read())

or

foo_codec(bar_codec('data', errors='ignore'), errors='replace')

and the user should never have to worry about whether something is
"encoding" or "decoding".

It should also be possible to look codecs up by name. For reversible
str<->str and bytes<->bytes transforms, the convention is that 'unfoo'
undoes 'foo' (including transforms that undo themselves). For
irreversible transforms, this is N/A. For text encodings (or other
str<->bytes transforms), the direction of transformation is inferred
from the input type.